Avoid being solely limited to U.K.
This may be paranoid, but if we were to work with a non U.K. Event, there may be future difficulties if we tried to take on contracts or other responsibilities to assist that event.  I am thinking of the African conference that was run a week or two before last year.  but technically we might find it hard to do anything on the Isle of Man conference (pymanx)

Just my paranoia showing

Not tested changes - on iPhone.